+++ lib/flutter_firebase_auth_ui.dart
@@ -1,7 +1,5 @@
 library flutter_firebase_auth_ui;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/typedef.dart';
+export 'src/signin_email_password_page.dart';
+export 'src/signin_phone_page.dart';